Carry RWI postings as domain values from crawler to node
The crawler→node chunk now carries []yacymodel.RWIPosting: the page
representation, posting chunk, and JSON envelope hold domain postings,
the crawler builds them directly, and the wire→domain shim in the node's
ingest consumer is gone.

RWIPosting's count and position fields become int, dropping the wire
column widths that had leaked into the domain type. The narrowing that
those widths forced is gone with them: the crawler assigns plain counts,
the 0x02 disk codec stores them as lossless varints, and only the YaCy
DHT wire form still speaks fixed-width columns, at its own edge.

+++ libraries/yacymodel/rwi_posting.go
@@ -6,18 +6,18 @@ type RWIPosting struct {
 	WordHash               Hash
 	URLHash                URLHash
 	LastModified           MicroDate
-	TitleWords             uint8
-	TextWords              uint16
-	Phrases                uint16
+	TitleWords             int
+	TextWords              int
+	Phrases                int
 	DocumentType           DocumentType
 	Language               Language
-	LocalLinks             uint8
-	ExternalLinks          uint8
-	URLLength              uint8
-	URLComponents          uint8
+	LocalLinks             int
+	ExternalLinks          int
+	URLLength              int
+	URLComponents          int
 	Appearance             Appearance
-	Hits                   uint8
-	TextPosition           uint16
-	PhraseRelativePosition uint8
-	PhrasePosition         uint8
+	Hits                   int
+	TextPosition           int
+	PhraseRelativePosition int
+	PhrasePosition         int
 }
